footer terminado y parte de resposing
