+++ package.json
@@ -24,6 +24,7 @@
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "material-icons": "^1.12.2",
     "rxjs": "~7.5.0",
+    "triggered-browser-animations": "^0.3.1",
     "tslib": "^2.3.0",
     "zone.js": "~0.12.0"
   },
