+++ src/app/app.component.html
@@ -1,2 +1,4 @@
-<app-nav></app-nav>
-<app-footer></app-footer>
+<main class="main">
+  <app-nav></app-nav>
+  <app-footer></app-footer>
+</main>
