+++ src/app/app.component.scss
@@ -1,0 +1,6 @@
+.main{
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
