--- src/app/componentes/footer/footer.component.html
@@ -1,8 +1,0 @@
-<div class="contenedor">
-  <h1>ksdj</h1>
-  <h2>ajbd</h2>
-
-  <div class="iconos">
-
-  </div>
-</div>
