--- src/app/componentes/footer/footer.component.scss
@@ -1,5 +1,0 @@
-.contenedor{
-  width: 100%;
-  height: 100%;
-  background-color:  rgb(183, 202, 219);
-}
