--- src/app/componentes/footer/footer.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import
-
-@Component({
-  selector: 'app-footer',
-  templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.scss']
-})
-export class FooterComponent {
-
-}
