--- src/app/componentes/nav/nav.component.scss
@@ -1,60 +1,0 @@
-.contenedor{
-  display: flex;
-  width: 100%;
-  height: 100%;
-  background-color: rgb(218, 184, 139);
-}
-
-.contenedor2{
-  display: flex;
-  justify-content: start;
-  width: 100%;
-}
-
-.contenedor3{
-  display: flex;
-  justify-content: space-around;
-  width: 100%;
-}
-
-.icono{
-  font-size: 30px;
-  margin-top: 2%;
-  color: rgb(255 255 255);
-}
-
-.titulos1{
-  margin-left: 2%;
-  color: rgb(0, 0, 0);
-  font-size: 29px;
-}
-
-.titulos2{
-  color: rgb(9, 9, 9);
-  margin-top: 1%;
-}
-
-.titulos3{
-  color: rgb(12, 12, 12);
-  margin-top: 1%;
-
-}
-.titulos4{
-
-  color: rgb(0, 0, 0);
-  margin-top: 1%;
-}
-.titulos5{
-  color: rgb(6, 6, 6);
-  margin-top: 1%;
-}
-
-.usuario{
-  color: rgb(255 255 255);
-  font-size: 24px;
-}
-
-.carrito{
-  color: rgb(255 255 255);
-  font-size: 24px;
-}
