+++ src/app/footer/footer.component.html
@@ -1,0 +1,25 @@
+<footer class="contenedor">
+
+
+  <div class="redes">
+    <img class="img-instragram" [src]="instagram">
+    <img class="img-Tikto" [src]="tikTok">
+    <img class="img-facebook" [src]="facebook">
+    <img class="img-youtube" [src]="youtube">
+  </div>
+
+
+  <div class="contenedor2">
+    <span  class="estilo">Privacy Policy</span>
+    <span class="estilo2">Terms of Use</span>
+    <span class="estilo3">@2022 HIKI</span>
+  </div>
+
+  <div class="contenedor3">
+    <span class="estilo2">@2022 arfa, Inc. All rights reversed. </span>
+    <span class="estilo">Don`t want to receive the HIKI email?</span>
+  </div>
+</footer>
+
+
+
