+++ src/app/nav/nav.component.html
@@ -1,25 +1,23 @@
-
 <div>
     <mat-toolbar class="contenedor">
-
-
     <div class="contenedor2">
         <fa-icon class="icono" [icon]="tienda"></fa-icon>
         <h1 class="titulos1">Store</h1>
     </div>
 
     <div class="contenedor3">
-      <span class="titulos2">HOME</span>
+      <!-- <span class="titulos2">HOME</span>
       <span class="titulos3">ABOUT US</span>
       <span class="titulos4">SHOP</span>
-      <span class="titulos5">CONTACT</span>
+      <span class="titulos5">CONTACT</span> -->
       <fa-icon class="usuario" [icon]="usuario"></fa-icon>
       <fa-icon class="carrito" [icon]="carrito"></fa-icon>
+      <!-- <fa-icon class="carrito" [icon]="barra"></fa-icon> -->
       <span class="example-spacer"></span>
 
     </div>
-  </mat-toolbar>
 
+  </mat-toolbar>
 
 </div>
