+++ src/app/nav/nav.component.scss
@@ -1,0 +1,268 @@
+.contenedor{
+  display: flex;
+  width: 100%;
+  height: 100%;
+  background-color: rgb(121, 180, 183);
+}
+
+.contenedor2{
+  display: flex;
+  justify-content: start;
+  width: 100%;
+}
+
+.contenedor3{
+  display: flex;
+  justify-content: space-around;
+  width: 100%;
+}
+
+.icono{
+  font-size: 30px;
+  margin-top: 2%;
+  color: rgb(2, 1, 1);
+}
+
+.titulos1{
+  margin-left: 2%;
+  color: rgb(0, 0, 0);
+  font-size: 29px;
+}
+
+.titulos2{
+  color: rgb(9, 9, 9);
+  margin-top: 1%;
+}
+
+.titulos3{
+  color: rgb(12, 12, 12);
+  margin-top: 1%;
+
+}
+.titulos4{
+
+  color: rgb(0, 0, 0);
+  margin-top: 1%;
+}
+.titulos5{
+  color: rgb(6, 6, 6);
+  margin-top: 1%;
+}
+
+.usuario{
+  color: rgb(0, 0, 0);
+  font-size: 24px;
+}
+
+.carrito{
+  color: rgb(0, 0, 0);
+  font-size: 24px;
+}
+
+.barra{
+  margin: 10%;
+}
+
+@media only screen and (max-width: 820px) and (min-width: 501px) {
+
+  .contenedor{
+    display: flex;
+    width: 100%;
+    height: 100%;
+    background-color: rgb(121, 180, 183);
+  }
+
+  .contenedor2{
+    display: flex;
+    justify-content: start;
+    width: 100%;
+  }
+
+  .contenedor3{
+    display: flex;
+    justify-content: space-around;
+    width: 100%;
+  }
+
+  .icono{
+    font-size: 30px;
+    margin-top: 2%;
+    color: rgb(2, 1, 1);
+  }
+
+  .titulos1{
+    margin-left: 2%;
+    color: rgb(0, 0, 0);
+    font-size: 29px;
+  }
+
+  .titulos2{
+    color: rgb(9, 9, 9);
+    margin-top: 1%;
+  }
+
+  .titulos3{
+    color: rgb(12, 12, 12);
+    margin-top: 1%;
+
+  }
+  .titulos4{
+
+    color: rgb(0, 0, 0);
+    margin-top: 1%;
+  }
+  .titulos5{
+    color: rgb(6, 6, 6);
+    margin-top: 1%;
+  }
+
+  .usuario{
+    color: rgb(0, 0, 0);
+    font-size: 24px;
+  }
+
+  .carrito{
+    color: rgb(0, 0, 0);
+    font-size: 24px;
+  }
+
+
+}
+
+@media only screen and (max-width: 500px) and (min-width: 341px) {
+  .contenedor{
+    display: flex;
+    width: 100%;
+    height: 100%;
+    background-color: rgb(121, 180, 183);
+  }
+
+  .contenedor2{
+    display: flex;
+    justify-content: start;
+    width: 100%;
+  }
+
+  .contenedor3{
+    display: flex;
+    justify-content: space-around;
+    width: 100%;
+  }
+
+  .icono{
+    font-size: 21px;
+    margin-top: 12%;
+    color: rgb(0, 0, 0);
+  }
+
+  .barra{
+    margin-left: 10%;
+    margin-top: 5%;
+  }
+
+  .titulos1{
+    margin-left: 2%;
+    color: rgb(0, 0, 0);
+    font-size: 25px;
+  }
+
+  .titulos2{
+    color: rgb(9, 9, 9);
+    margin-top: 1%;
+  }
+
+  .titulos3{
+    color: rgb(12, 12, 12);
+    margin-top: 1%;
+
+  }
+  .titulos4{
+
+    color: rgb(0, 0, 0);
+    margin-top: 1%;
+  }
+  .titulos5{
+    color: rgb(6, 6, 6);
+    margin-top: 1%;
+  }
+
+  .usuario{
+    color: rgb(0, 0, 0);
+    font-size: 24px;
+    display: flex;
+    flex-wrap: nowrap;
+  }
+
+  .carrito{
+    color: rgb(0, 0, 0);
+    font-size: 24px;
+    display: flex;
+    flex-wrap: nowrap;
+  }
+
+}
+
+@media only screen and (max-width: 820px) and (min-width: 501px) {
+  .contenedor{
+    display: flex;
+    width: 100%;
+    height: 100%;
+    background-color: rgb(121, 180, 183);
+  }
+
+  .contenedor2{
+    display: flex;
+    justify-content: start;
+    width: 100%;
+  }
+
+  .contenedor3{
+    display: flex;
+    justify-content: space-around;
+    width: 100%;
+  }
+
+  .icono{
+    font-size: 30px;
+    margin-top: 2%;
+    color: rgb(2, 1, 1);
+  }
+
+  .titulos1{
+    margin-left: 2%;
+    color: rgb(0, 0, 0);
+    font-size: 29px;
+  }
+
+  .titulos2{
+    color: rgb(9, 9, 9);
+    margin-top: 1%;
+  }
+
+  .titulos3{
+    color: rgb(12, 12, 12);
+    margin-top: 1%;
+
+  }
+  .titulos4{
+
+    color: rgb(0, 0, 0);
+    margin-top: 1%;
+  }
+  .titulos5{
+    color: rgb(6, 6, 6);
+    margin-top: 1%;
+  }
+
+  .usuario{
+    color: rgb(0, 0, 0);
+    font-size: 24px;
+  }
+
+  .carrito{
+    color: rgb(0, 0, 0);
+    font-size: 24px;
+  }
+
+
+}
