+++ src/app/nav/nav.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { faUser } from '@fortawesome/free-solid-svg-icons';
 import { faCartShopping } from '@fortawesome/free-solid-svg-icons';
-import { faThList } from '@fortawesome/free-solid-svg-icons';
 import { faStore } from '@fortawesome/free-solid-svg-icons';
+import { faBars } from '@fortawesome/free-solid-svg-icons';
 @Component({
   selector: 'app-nav',
   templateUrl: './nav.component.html',
@@ -12,4 +12,5 @@ export class NavComponent {
   usuario = faUser;
   carrito = faCartShopping;
   tienda= faStore;
+  barra = faBars;
 }
