+++ src/styles.scss
@@ -1,4 +1,5 @@
 /* You can add global styles to this file, and also import other style files */
 body{
-  background: rgb(243, 233, 221);  /* fallback for old browsers */
+  background:rgb(254, 251, 243);/* fallback for old browsers */
+  margin: 0%;
 }
